+++ src/main/java/de/retest/web/RecheckSeleniumAdapter.java
@@ -10,6 +10,7 @@ import java.nio.charset.StandardCharsets;
 import java.util.Collections;
 import java.util.Map;
 import java.util.Set;
+import java.util.stream.Stream;
 
 import org.apache.commons.io.IOUtils;
 import org.openqa.selenium.JavascriptExecutor;
@@ -22,13 +23,18 @@ import org.slf4j.LoggerFactory;
 import de.retest.recheck.RecheckAdapter;
 import de.retest.recheck.RecheckOptions;
 import de.retest.recheck.meta.MetadataProvider;
+import de.retest.recheck.report.ActionReplayResult;
 import de.retest.recheck.ui.DefaultValueFinder;
 import de.retest.recheck.ui.descriptors.RootElement;
 import de.retest.recheck.ui.descriptors.idproviders.RetestIdProvider;
+import de.retest.recheck.ui.diff.ElementDifference;
 import de.retest.web.mapping.PathsToWebDataMapping;
 import de.retest.web.meta.SeleniumMetadataProvider;
 import de.retest.web.screenshot.ScreenshotProvider;
+import de.retest.web.selenium.AutocheckingRecheckDriver;
+import de.retest.web.selenium.ImplicitDriverWrapper;
 import de.retest.web.selenium.UnbreakableDriver;
+import de.retest.web.selenium.WriteToResultWarningConsumer;
 import de.retest.web.util.SeleniumWrapperUtil;
 import de.retest.web.util.SeleniumWrapperUtil.WrapperOf;
 
@@ -42,6 +48,8 @@ public class RecheckSeleniumAdapter implements RecheckAdapter {
 
 	private final RetestIdProvider retestIdProvider;
 	private final ScreenshotProvider screenshotProvider;
+
+	private ActionReplayResult actionReplayResult;
 
 	public RecheckSeleniumAdapter( final RecheckOptions options ) {
 		retestIdProvider = options.getRetestIdProvider();
@@ -85,22 +93,53 @@ public class RecheckSeleniumAdapter implements RecheckAdapter {
 
 	@Override
 	public Set<RootElement> convert( final Object toVerify ) {
-		if ( SeleniumWrapperUtil.isWrapper( WrapperOf.ELEMENT, toVerify ) ) {
-			return convert( SeleniumWrapperUtil.getWrapped( WrapperOf.ELEMENT, toVerify ) );
+		final RemoteWebElement webElement = retrieveWebElement( toVerify );
+		if ( webElement != null ) {
+			return convertWebElement( webElement );
 		}
-		if ( toVerify instanceof RemoteWebElement ) {
-			return convertWebElement( (RemoteWebElement) toVerify );
-		}
-		if ( toVerify instanceof UnbreakableDriver ) {
-			return convertWebDriver( (UnbreakableDriver) toVerify );
-		}
-		if ( SeleniumWrapperUtil.isWrapper( WrapperOf.DRIVER, toVerify ) ) {
-			return convert( SeleniumWrapperUtil.getWrapped( WrapperOf.DRIVER, toVerify ) );
-		}
-		if ( toVerify instanceof RemoteWebDriver ) {
-			return convertWebDriver( (RemoteWebDriver) toVerify );
+		final WebDriver webDriver = retrieveWebDriver( unwrapImplicitDriver( toVerify ) );
+		if ( webDriver != null ) {
+			return convertWebDriver( webDriver );
 		}
 		throw new IllegalArgumentException( "Cannot convert objects of type '" + toVerify.getClass().getName() + "'." );
+	}
+
+	private RemoteWebElement retrieveWebElement( final Object toVerify ) {
+		if ( SeleniumWrapperUtil.isWrapper( WrapperOf.ELEMENT, toVerify ) ) {
+			return retrieveWebElement( SeleniumWrapperUtil.getWrapped( WrapperOf.ELEMENT, toVerify ) );
+		}
+		if ( toVerify instanceof RemoteWebElement ) {
+			return (RemoteWebElement) toVerify;
+		}
+		return null;
+	}
+
+	private Object unwrapImplicitDriver( final Object toVerify ) {
+		if ( toVerify instanceof AutocheckingRecheckDriver ) {
+			throw new UnsupportedOperationException( String.format(
+					"The '%s' does implicit checking after each action, " // 
+							+ "therefore no explicit check with 'Recheck#check' is needed. " // 
+							+ "Please remove the explicit check. " //
+							+ "For more information, please have a look at https://docs.retest.de/recheck-web/introduction/usage/.",
+					toVerify.getClass().getSimpleName() ) );
+		}
+		if ( toVerify instanceof ImplicitDriverWrapper ) {
+			return ((ImplicitDriverWrapper) toVerify).getWrappedDriver();
+		}
+		return toVerify;
+	}
+
+	private WebDriver retrieveWebDriver( final Object toVerify ) {
+		if ( toVerify instanceof UnbreakableDriver ) {
+			return (UnbreakableDriver) toVerify;
+		}
+		if ( SeleniumWrapperUtil.isWrapper( WrapperOf.DRIVER, toVerify ) ) {
+			return retrieveWebDriver( SeleniumWrapperUtil.getWrapped( WrapperOf.DRIVER, toVerify ) );
+		}
+		if ( toVerify instanceof RemoteWebDriver ) {
+			return (RemoteWebDriver) toVerify;
+		}
+		return null;
 	}
 
 	Set<RootElement> convertWebDriver( final WebDriver driver ) {
@@ -127,6 +166,8 @@ public class RecheckSeleniumAdapter implements RecheckAdapter {
 
 		if ( driver instanceof UnbreakableDriver ) {
 			((UnbreakableDriver) driver).setLastActualState( lastChecked );
+			((UnbreakableDriver) driver)
+					.setWarningConsumer( new WriteToResultWarningConsumer( this::retrieveDifferences ) );
 		}
 
 		return Collections.singleton( lastChecked );
@@ -149,6 +190,13 @@ public class RecheckSeleniumAdapter implements RecheckAdapter {
 		}
 	}
 
+	private Stream<ElementDifference> retrieveDifferences() {
+		if ( actionReplayResult == null ) {
+			return Stream.empty();
+		}
+		return actionReplayResult.getAllElementDifferences().stream();
+	}
+
 	@Override
 	public Map<String, String> retrieveMetadata( final Object toCheck ) {
 		final MetadataProvider provider = SeleniumMetadataProvider.of( toCheck );
@@ -160,4 +208,8 @@ public class RecheckSeleniumAdapter implements RecheckAdapter {
 		return defaultValueFinder;
 	}
 
+	@Override
+	public void notifyAboutDifferences( final ActionReplayResult actionReplayResult ) {
+		this.actionReplayResult = actionReplayResult;
+	}
 }