+++ src/main/java/de/retest/web/screenshot/ScreenshotProviders.java
@@ -11,10 +11,15 @@ import org.openqa.selenium.WebElement;
 
 import com.assertthat.selenium_shutterbug.core.Shutterbug;
 
+import de.retest.web.RecheckWebOptions;
+import de.retest.web.RecheckWebProperties;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 public class ScreenshotProviders {
+	
+	public static final ScreenshotProvider DEFAULT = new ViewportOnlyMinimalScreenshot();
+	public static final ScreenshotProvider NONE = new NoScreenshot();
 
 	public static class ScreenshotProviderConverter implements Converter<ScreenshotProvider> {
 
@@ -26,11 +31,14 @@ public class ScreenshotProviders {
 				case "viewportOnly":
 					return new ViewportOnlyMinimalScreenshot();
 				case "none":
-					return new NoScreenshot();
+					log.info( "ScreenshotProvider has been set to 'none' either via property "
+							+ RecheckWebProperties.SCREENSHOT_PROVIDER_PROPERTY_KEY + " or via "
+							+ RecheckWebOptions.class.getSimpleName() + ", will create NO screenshots." );
+					return NONE;
 				default:
 					log.warn( "Unknown configured screenshot provider '{}'. Using default value 'viewportOnly'.",
 							input );
-					return new ViewportOnlyMinimalScreenshot();
+					return DEFAULT;
 			}
 		}
 
@@ -45,10 +53,13 @@ public class ScreenshotProviders {
 	public static BufferedImage shoot( final WebDriver driver, final WebElement element,
 			final ScreenshotProvider screenshotProvider ) {
 		try {
+			final long startTime = System.currentTimeMillis();
 			if ( element != null ) {
 				return shootElement( driver, element );
 			}
-			return screenshotProvider.shoot( driver );
+			final BufferedImage result = screenshotProvider.shoot( driver );
+			log.info( "Took {}ms to create the screenshot.", System.currentTimeMillis() - startTime );
+			return result;
 		} catch ( final Exception e ) {
 			log.error( "Exception creating screenshot for check.", e );
 			return null;